Rename TreeView -> EntityTree, TreePresenter -> VirtualizedTree
