+++ src/lib/containers/entity_finder/EntityFinder.tsx
@@ -31,7 +31,7 @@ import {
 } from './details/EntityDetailsList'
 import { SelectionPane } from './SelectionPane'
 import { EntityTreeNodeType } from './tree/VirtualizedTree'
-import { FinderScope, TreeView } from './tree/TreeView'
+import { FinderScope, EntityTree } from './tree/EntityTree'
 import pluralize from 'pluralize'
 import { entityTypeToFriendlyName } from '../../utils/functions/EntityTypeUtils'
 
@@ -356,7 +356,7 @@ export const EntityFinder: React.FunctionComponent<EntityFinderProps> = ({
             <div style={searchActive ? { display: 'none' } : {}}>
               {treeOnly ? (
                 <div>
-                  <TreeView
+                  <EntityTree
                     toggleSelection={toggleSelection}
                     showDropdown={true}
                     visibleTypes={selectableAndVisibleTypesInTree}
@@ -382,7 +382,7 @@ export const EntityFinder: React.FunctionComponent<EntityFinderProps> = ({
                           className="TreeViewReflexElement"
                           flex={0.24}
                         >
-                          <TreeView
+                          <EntityTree
                             selectedEntities={selectedEntities}
                             setDetailsViewConfiguration={setConfigFromTreeView}
                             showDropdown={true}
